static class variables should be of type double

+++ @sgem/sgem.m
@@ -373,7 +373,7 @@ classdef sgem < handle
                 if precision < 1
                     error('The precision need to be larger or equal to 1    ');
                 end
-                precision = newValue;
+                precision = double(newValue);
                 % We call the mex interface to make this the default
                 % working precision
                 sgem_mex('setWorkingPrecision', precision);
@@ -389,7 +389,7 @@ classdef sgem < handle
             end
             if nargin >= 1
                 if newValue >= 1
-                    precision = newValue;
+                    precision = double(newValue);
                 else
                     % This means plotting as many nonzero digits as
                     % available
@@ -408,7 +408,7 @@ classdef sgem < handle
             end
             if nargin >= 1
                 if (newValue == 1) || (newValue == 0)
-                    likeMatlab = newValue;
+                    likeMatlab = double(newValue);
                 end
             end
             value = likeMatlab;
